refactor: Se cambian tipos de datos

+++ src/main/java/co/com/ies/smolplus/context/moduledevicemanager/application/DeviceManagerService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface DeviceManagerService {
 
-    List<DeviceEstablishmentDTO> getDeviceEstablishmentByEstablishmentId(Integer id);
+    List<DeviceEstablishmentDTO> getDeviceEstablishmentByEstablishmentId(Long id);
 }
